server/requirements-dev.txt
# Specifies only dev-specific requirements
# But imports the runtime ones too
-r requirements.txt

pytest
pytest-cov
pytest-mock
flake8
flake8-bugbear
black
lxml-stubs